SAGE/requirements.txt
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124
transformers==4.57.0
peft==0.18.0
timm==1.0.26
ultralytics==8.4.14
opencv-python==4.12.0.88
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
paddlepaddle==3.3.0
paddleocr==3.4.0
paddlex==3.4.2
decord==0.6.0
scikit-image==0.25.2
shapely==2.1.2
pytesseract==0.3.13
librosa==0.11.0
openai-whisper==20250625
soundfile==0.13.1
numpy==2.2.6
pandas==2.3.3
scipy==1.16.3
scikit-learn==1.7.2
tqdm==4.67.1
pyyaml==6.0.2